all: kickstart project
This commit start exploring how to build application using wire
dependency injection base. My assumption is we can prepare a set
that we uses commonly in our project such as http server.

The design is we create interface called App to provides functionality
to start and stop application and make a specific protocol application
with this protocol.